pyproject.toml
[project]
name = "joa-qartod-config"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Liz Dobbins", email = "liz.dobbins4@gmail.com" }
]
requires-python = ">=3.12"
dependencies = [
    "ioos-qc>=2.3.0",
    "jupyter-bokeh>=4.0.5",
    "mysql-connector-python>=9.5.0",
    "pandas>=2.3.3",
    "python-dotenv>=1.2.1",
    "simplejson>=3.20.2",
    "sqlalchemy>=2.0.45",
]

[project.scripts]
joa-qartod-config = "joa_qartod_config:main"

[project.optional-dependencies]
testing = [
    "pytest>=9.0.2",
]

[build-system]
requires = ["uv_build>=0.8.22,<0.9.0"]
build-backend = "uv_build"

[tool.pytest.ini_options]
# from https://learn.scientific-python.org/development/guides/pytest/
minversion = "6.0"
addopts = ["-ra", "--showlocals", "--strict-markers", "--strict-config"]
xfail_strict = true
filterwarnings = ["error"]
log_cli_level = "INFO"
testpaths = [
  "tests",
]

[tool.ruff.lint]
extend-select = [
  "ARG",      # flake8-unused-arguments
  "B",        # flake8-bugbear
  "C4",       # flake8-comprehensions
  "EM",       # flake8-errmsg
  "EXE",      # flake8-executable
  "FURB",     # refurb
  "G",        # flake8-logging-format
  "I",        # isort
  "ICN",      # flake8-import-conventions
  "NPY",      # NumPy specific rules
  "PD",       # pandas-vet
  "PGH",      # pygrep-hooks
  "PIE",      # flake8-pie
  "PL",       # pylint
  "PT",       # flake8-pytest-style
  "PTH",      # flake8-use-pathlib
  "PYI",      # flake8-pyi
  "RET",      # flake8-return
  "RUF",      # Ruff-specific
  "SIM",      # flake8-simplify
  "T20",      # flake8-print
  "UP",       # pyupgrade
  "YTT",      # flake8-2020
]
ignore = [
  "ISC001",   # Conflicts with formatter
  "PLR09",    # Too many <...>
  "PLR2004",  # Magic value used in comparison
]
typing-modules = ["mypackage._compat.typing"]
isort.required-imports = ["from __future__ import annotations"]

[tool.ruff.lint.per-file-ignores]
"tests/**" = ["T20"]

[tool.ruff]
# Always generate Python 3.12-compatible code.
target-version = "py312"
